requirements.txt
altgraph==0.17.5
macholib==1.16.4
packaging==25.0
pillow==12.1.0
pyinstaller==6.17.0
pyinstaller-hooks-contrib==2025.11
setuptools==80.9.0
